revised implicit primitive type converstion; configuration of cling-tidy cheks

+++ src/window.cpp
@@ -6,7 +6,7 @@
 void window::hanning(SAMPLE_ARRAY &data)
 {
     auto func = [size = data.size(), index = 0](SAMPLE &sample) mutable {
-        sample *= 0.5 - 0.5 * std::cos(2 * M_PI * index++ / (size - 1));
+        sample *= 0.5 - 0.5 * std::cos(2 * M_PI * index++ / static_cast<double>(size - 1));
     };
 
     std::for_each(data.begin(), data.end(), func);
@@ -15,7 +15,7 @@ void window::hanning(SAMPLE_ARRAY &data)
 void window::hamming(SAMPLE_ARRAY &data)
 {
     auto func = [size = data.size(), index = 0](SAMPLE &sample) mutable {
-        sample *= 0.54 - 0.46 * std::cos(2 * M_PI * index++ / (size - 1));
+        sample *= 0.54 - 0.46 * std::cos(2 * M_PI * index++ / static_cast<double>(size - 1));
     };
 
     std::for_each(data.begin(), data.end(), func);
@@ -25,8 +25,8 @@ void window::blackman(SAMPLE_ARRAY &data)
 {
     auto func = [size = data.size(), index = 0](SAMPLE &sample) mutable {
         sample *= 0.42
-            - 0.5 * std::cos(2 * M_PI * index / (size - 1))
-            + 0.08 * std::cos(4 * M_PI * index / (size - 1));
+            - 0.5 * std::cos(2 * M_PI * index / static_cast<double>(size - 1))
+            + 0.08 * std::cos(4 * M_PI * index / static_cast<double>(size - 1));
         ++index;
     };
 
@@ -37,10 +37,10 @@ void window::flattop(SAMPLE_ARRAY &data)
 {
     auto func = [size = data.size(), index = 0](SAMPLE &sample) mutable {
         sample *= 0.21557895
-            - 0.41663158 * std::cos(2 * M_PI * index / (size - 1))
-            + 0.277263158 * std::cos(4 * M_PI * index / (size - 1))
-            - 0.083578947 * std::cos(6 * M_PI * index / (size - 1))
-            + 0.006947368 * std::cos(8 * M_PI * index / (size - 1));
+            - 0.41663158 * std::cos(2 * M_PI * index / static_cast<double>(size - 1))
+            + 0.277263158 * std::cos(4 * M_PI * index / static_cast<double>(size - 1))
+            - 0.083578947 * std::cos(6 * M_PI * index / static_cast<double>(size - 1))
+            + 0.006947368 * std::cos(8 * M_PI * index / static_cast<double>(size - 1));
         ++index;
     };
